+++ gpu-parallel/v1/correct/bcsstk16_test.c
@@ -3,17 +3,17 @@
 #include<assert.h>
 #include<time.h>
 #include"../src/cg_gpu_parallel.h"
-#include"../../matrices/matrix_load.h"
+#include"../../matrices/matrix_load_gpu.h"
 
 #define epsilon 1E-9
 
 int main()
 {
     // BCSSTK16
-    float* cg_solution;
-    float* FFGE_matrix;
-    float b_16[4884];
-    float init_16[4884];
+    double* cg_solution;
+    double* FFGE_matrix;
+    double b_16[4884];
+    double init_16[4884];
     float startTime;
     float endTime;
     float timeElapsed;
